fix: Generate proper URLs in llms.txt
The plugin used before does not appear to be in GitHub or
actively maintained. It had a bug where all URLs generated would
be of the form `https:/adcontextprotocl.org` which missing a slash.

Switch to using the llmstxt.org listed docusaurus plugin.
This does change the generated file to include more (truncated)
descriptions, but the URLs are correct.