Skip changeset check on Version Packages PR
The changeset-release/main PR is created by the changesets automation
and should not be required to have a changeset of its own. That would
defeat the purpose of the Version Packages workflow.

This adds a conditional to skip the changeset check when the PR is
from the changeset-release/main branch.